requirements.txt
# ─── Core Framework ───────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3

# ─── LangChain Stack ──────────────────────────────────────────
langchain==0.3.9
langchain-core==0.3.21
langchain-anthropic==0.3.0
langchain-openai==0.2.11          # OpenAI embeddings only
langchain-community==0.3.9
langgraph==0.2.56

# ─── LangSmith ────────────────────────────────────────────────
langsmith==0.1.147

# ─── RAG / Vector Store ───────────────────────────────────────
pinecone==6.0.0
langchain-pinecone==0.2.13
tiktoken==0.8.0

# ─── HTTP & API ───────────────────────────────────────────────
httpx==0.27.2
python-dotenv==1.0.1

# ─── Data ─────────────────────────────────────────────────────
pandas==2.2.3

# ─── Dev / Testing ────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0